requirements.txt
streamlit
huggingface-hub
pillow
python-dotenv
